+++ knble/src/main/java/ovh/karewan/knble/interfaces/BleNotifyCallback.java
@@ -1,0 +1,8 @@
+package ovh.karewan.knble.interfaces;
+
+public interface BleNotifyCallback {
+	void onNotifyEnabled();
+	void onNotifyDisabled();
+	void onNotify(byte[] data);
+	void onNotifyFailed();
+}